backend/dukaan/requirements.txt
appdirs==1.4.4
asgiref==3.8.1
attrs==24.2.0
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
colorama==0.4.6
cssselect2==0.7.0
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
fonttools==4.53.1
h11==0.14.0
html5lib==1.1
idna==3.10
importlib_metadata==8.5.0
mysqlclient==2.2.4
outcome==1.3.0.post0
packaging==24.1
pdfkit==1.0.0
pillow==10.4.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
pydyf==0.11.0
pyee==11.1.1
PyJWT==2.9.0
pymongo==4.8.0
pyphen==0.16.0
pyppeteer==2.0.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
tinycss2==1.3.0
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==1.26.20
weasyprint==62.3
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==10.4
wsproto==1.2.0
zipp==3.20.2
zopfli==0.2.3
